Stars.jl Documentation

Stars.jl Documentation

Index

Stars.GeoArray — Type

GeoArray(A::AbstractArray{<:RealOrMissing}, f::AffineMap = default_affinemap, crs::WellKnownText = crs2wkt(""))
GeoArray(A::AbstractArray{<:RealOrMissing}, f::AffineMap, crs::AbstractString)
GeoArray(A::AbstractArray{<:RealOrMissing}, x::AbstractRange, y::AbstractRange, args...)

GeoArray(fn::AbstractString, bands = nothing) 
GeoArray(ga::AbstractGeoArray; vals::AbstractArray{<:RealOrMissing})

Construct a GeoArray from any Array. Meanwhile, AffineMap and CRS will be generated.

Arguments

A: An AbstractArray, at least 2d Array

Examples

ga = GeoArray(rand(10,10,1))
GeoArray(rand(10,10))

GeoArray(ga, vals = ga.A)                        # copy `f` and `crs` from `ga`
GeoArray(rand(10,10), 1:10, 2:11)                # define coordinate by `x` and `y`
GeoArray(rand(7, 4), bbox([70, 15, 140, 55]...)) # define coordinate by `bbox`

# 4-d array
ga = GeoArray(rand(10, 10, 2, 3))
GeoArray(ga, vals = ga.A)

Base.getindex — Function

getindex(ga::AbstractGeoArray, i::AbstractRange, j::AbstractRange, k::TYPE_index = :)
TYPE_index = Union{Colon,AbstractRange,Integer}

Index a AbstractGeoArray with AbstractRanges to get a cropped AbstractGeoArray with the correct AffineMap set.

Examples

julia> ga[2:3,2:3,1]
2x2x1 Array{Float64, 3} with AffineMap([1.0 0.0; 0.0 1.0], [1.0, 1.0]) and undefined CRS

Stars.cast_to_gdal — Method

Converts type of Array for one that exists in GDAL.

Stars.cellIJ — Method

cellIJ(r::GeoArray, point::Tuple)

@return

Ilon, Ilat

Stars.clip_point — Method

clip_point(r, shp; ngrid = 10)

shp: DataFrame with the column of geom

Stars.equals — Method

Check whether two GeoArrayss a and b are geographically equal, although not necessarily in content.

Stars.flipud! — Method

Function to flip GeoArray upside down to adjust to GDAL ecosystem.

Stars.gdal_polygonize — Function

gdal_polygonize(raster_file, band = 1, out_file = "out.shp"; 
    fieldname = "grid", nodata = NaN)

Creates vector polygons for all connected regions of pixels in the raster sharing a common pixel value.

References

https://gdal.org/programs/gdal_polygonize.html

Stars.gdalinfo — Method

gdalinfo(ga::GeoArray)
gdalinfo(file::AbstractString)

get detailed GDAL information

return

range : [lonmin, lonmax, latmin, latmax]

cellsize : [dx, dy]

coords : [lon, lat]

dim : (width, height, ntime)

Stars.get_nodata — Method

Retrieves nodata value from RasterBand.

Stars.is_rotated — Method

Check wether the AffineMap of a AbstractGeoArray contains rotations.

Stars.maskCoords — Method

dcoord = stas_sf(mask)

Stars.mask_flags — Method

Takes bitwise OR-ed set of status flags and returns flags.

Stars.readGDAL — Method

readGDAL(file::String, options...)
readGDAL(files::Array{String,1}, options)

Arguments:

options: other parameters to ArchGDAL.read(dataset, options...).

Return

Stars.shrink_bbox — Function

shrink_bbox(ga, mask::AbstractArray{Bool, 2} = nothing)

Only true values in mask will be kept.

Stars.st_as_sf — Function

st_as_sf(ga::AbstractGeoArray, mask::Union{Nothing, AbstractArray{Bool, 2}} = nothing)
st_as_sf(file::AbstractString, mask = nothing)

If the input is a file path, shrink_bbox will be applied.

Stars.st_as_stars — Method

st_as_stars(data::AbstractArray{T, 2}, mask::AbstractArray{Bool, 2})

st_as_stars(data::AbstractArray{T, 2}, r_mask::GeoArray; outfile = nothing)
st_as_stars(data::AbstractArray{T, 2}, mask::AbstractArray{Bool, 2}, 
    b::bbox; kw...)
st_as_stars(data::AbstractArray{T, 2}, mask::AbstractArray{Bool, 2}, 
    b::AbstractArray{<:Real, 1}; kw...)

Stars.st_bbox! — Method

Set geotransform of AbstractGeoArray by specifying a bounding box. Note that this only can result in a non-rotated or skewed AbstractGeoArray.

Stars.st_clip — Method

clipbbox(x::AbstractGeoArray, b::bbox) clipbbox(x::AbstractGeoArray, y::AbstractGeoArray)

Stars.st_coordinates! — Method

st_coordinates!(ga, x::AbstractUnitRange, y::AbstractUnitRange)

Set AffineMap of AbstractGeoArray by specifying the center coordinates for each x, y dimension by a UnitRange.

Stars.st_coordinates — Method

st_coordinates(ga)

st_coordinates is similar as st_dim, but for irregular AbstractGeoArray.

Stars.st_coords — Method

st_coords(ga; mid)

st_coords is for regular AbstractGeoArray.

Return

X,Y: A matrix with the dimension of [NX, NY]

Stars.st_crs! — Method

st_crs!(ga::AbstractGeoArray, crs::WellKnownText{GeoFormatTypes.CRS,<:AbstractString})
st_crs!(ga::AbstractGeoArray, crs::GFT.CoordinateReferenceSystemFormat)
st_crs!(ga::AbstractGeoArray, epsgcode::Int)

Set CRS on AbstractGeoArray by epsgcode or string

Stars.st_dim — Method

st_dim(ga::AbstractGeoArray; mid::Vector{Int} = [1, 1])
st_dim(b::bbox, cellsize::T) where {T <: Real}
st_dim(ga::AbstractGeoArray, dim::Symbol; mid::Vector{Int} = [1, 1])

Get dimensions of x and y

Stars.st_read — Function

st_read(fn::AbstractString, bands = nothing)

Return

A AbstractGeoArray object

Stars.st_write — Method

st_write(ga::AbstractGeoArray, fn::AbstractString; nodata = nothing)
st_write(A::AbstractArray{T}, b::bbox, fn::AbstractString;nodata = nothing)

Arguments

A: 2d or 3d array

fn: outfile path

nodata: should be in the same data type as ga.A (or A)

Stars.SpatialPixelsDataFrame — Type

SpatialPixelsDataFrame{
    T<:AbstractArray{<:Real, 2}, 
    C<:AbstractArray{<:Real, 2}, B<:bbox, P}  <: AbstractSpatialPoints

SpatialPixelsDataFrame(data::AbstractArray{<:Real, 2}, coords::AbstractArray{<:Real, 2}, b::bbox, proj = 4326)

Usage

SpatialPixelsDataFrame(data, coords, bbox, proj)

Stars.bbox — Type

bbox(xmin, ymin, xmax, ymax)
bbox(;xmin, ymin, xmax, ymax)
bbox2tuple(b::bbox)
bbox2vec(b::bbox)
bbox2affine(size::Tuple{Integer, Integer}, b::bbox)

Spatial bounding box

Index

Stars.GeoArray

Stars.SpatialPixelsDataFrame

Stars.bbox

Base.getindex

Stars.cast_to_gdal

Stars.cellIJ

Stars.clip_point

Stars.equals

Stars.flipud!

Stars.gdal_polygonize

Stars.gdalinfo

Stars.get_nodata

Stars.is_rotated

Stars.maskCoords

Stars.mask_flags

Stars.readGDAL

Stars.readGDAL

Stars.shrink_bbox

Stars.st_as_sf

Stars.st_as_stars

Stars.st_bbox!

Stars.st_clip

Stars.st_coordinates

Stars.st_coordinates!

Stars.st_coords

Stars.st_crs!

Stars.st_dim

Stars.st_read

Stars.st_read

Stars.st_write

Stars.st_write