Stars.jl Documentation

Stars.jl Documentation

Index

Stars.GeoArray — Type

GeoArray{T<:Union{Missing, Real}} <: AbstractArray{T, 3}
GeoArray(A::AbstractArray{T, 3})

Construct a GeoArray from any Array. A default AffineMap and CRS will be generated.

Examples

julia> GeoArray(rand(10,10,1))
10x10x1 Array{Float64, 3} with AffineMap([1.0 0.0; 0.0 1.0], [0.0, 0.0]) and undefined CRS

Base.getindex — Function

getindex(ga::GeoArray, i::AbstractRange, j::AbstractRange, k::Union{Colon,AbstractRange,Integer})

Index a GeoArray with AbstractRanges to get a cropped GeoArray with the correct AffineMap set.

Examples

julia> ga[2:3,2:3,1]
2x2x1 Array{Float64, 3} with AffineMap([1.0 0.0; 0.0 1.0], [1.0, 1.0]) and undefined CRS

Stars.cast_to_gdal — Method

Converts type of Array for one that exists in GDAL.

Stars.cellIJ — Method

cellIJ(r::GeoArray, point::Tuple)

@return

Ilon, Ilat

Stars.clip_point — Method

clip_point(r, shp; ngrid = 10)

shp: DataFrame with the column of geom

Stars.equals — Method

Check whether two GeoArrayss a and b are geographically equal, although not necessarily in content.

Stars.flipud! — Method

Function to flip GeoArray upside down to adjust to GDAL ecosystem.

Stars.gdal_polygonize — Function

gdal_polygonize(raster_file, band = 1, out_file = "out.shp"; 
    fieldname = "grid", nodata = NaN)

Creates vector polygons for all connected regions of pixels in the raster sharing a common pixel value.

References

https://gdal.org/programs/gdal_polygonize.html

Stars.gdalinfo — Method

gdalinfo(ga::GeoArray)
gdalinfo(file::AbstractString)

get detailed GDAL information

return

range : [lonmin, lonmax, latmin, latmax]

cellsize : [dx, dy]

coords : [lon, lat]

dim : (width, height, ntime)

Stars.get_nodata — Method

Retrieves nodata value from RasterBand.

Stars.maskCoords — Method

dcoord = stas_sf(mask)

Stars.mask_flags — Method

Takes bitwise OR-ed set of status flags and returns flags.

Stars.readGDAL — Method

readGDAL(file::String, options...)
readGDAL(files::Array{String,1}, options)

Arguments:

options: other parameters to ArchGDAL.read(dataset, options...).

Stars.shrink_bbox — Function

shrink_bbox(ga, mask::AbstractArray{Bool, 2} = nothing)

only true values in mask will be kept.

Stars.st_as_sf — Function

st_as_sf(ga::GeoArray, mask::Union{Nothing, AbstractArray{Bool, 2}} = nothing)
st_as_sf(file::AbstractString, mask = nothing)

If the input is a file path, shrink_bbox will be applied.

Stars.st_as_stars — Method

st_as_stars(data::AbstractArray{T, 2}, mask::AbstractArray{Bool, 2})

st_as_stars(data::AbstractArray{T, 2}, r_mask::GeoArray; outfile = nothing)
st_as_stars(data::AbstractArray{T, 2}, mask::AbstractArray{Bool, 2}, 
    b::bbox; kw...)
st_as_stars(data::AbstractArray{T, 2}, mask::AbstractArray{Bool, 2}, 
    b::AbstractArray{<:Real, 1}; kw...)

Stars.st_bbox! — Method

Set geotransform of GeoArray by specifying a bounding box. Note that this only can result in a non-rotated or skewed GeoArray.

Stars.st_clip — Method

clipbbox(x::GeoArray, b::bbox) clipbbox(x::GeoArray, y::GeoArray)

Stars.st_coordinates! — Method

st_coordinates!(ga, x::AbstractUnitRange, y::AbstractUnitRange)

Set AffineMap of GeoArray by specifying the center coordinates for each x, y dimension by a UnitRange.

Stars.st_coordinates — Method

st_coordinates(ga)

st_coordinates is similar as st_dim, but for irregular GeoArray.

Stars.st_coords — Method

st_coords(ga; mid)

st_coords is for regular GeoArray.

Return

X,Y: A matrix with the dimension of [NX, NY]

Stars.st_crs! — Method

st_crs!(ga::GeoArray, crs::WellKnownText{GeoFormatTypes.CRS,<:AbstractString})
st_crs!(ga::GeoArray, crs::GFT.CoordinateReferenceSystemFormat)
st_crs!(ga::GeoArray, epsgcode::Int)

Set CRS on GeoArray by epsgcode or string

Stars.st_dim — Method

st_dim(ga::GeoArray; mid::Vector{Int} = [1, 1])
st_dim(b::bbox, cellsize::T) where {T <: Real}
st_dim(ga::GeoArray, dim::Symbol; mid::Vector{Int} = [1, 1])

Get dimensions of x and y

Stars.st_read — Function

st_read(fn::AbstractString, bands = nothing)

Return

A GeoArray object

Stars.st_write — Method

st_write(ga::GeoArray, fn::AbstractString; nodata = nothing)
st_write(A::AbstractArray{T}, b::bbox, fn::AbstractString;nodata = nothing)

Arguments

A: 2d or 3d array

fn: outfile path

nodata: should be in the same data type as ga.A (or A)

Stars.SpatialPixelsDataFrame — Type

SpatialPixelsDataFrame{
    T<:AbstractArray{<:Real, 2}, 
    C<:AbstractArray{<:Real, 2}, B<:bbox, P}  <: AbstractSpatialPoints

SpatialPixelsDataFrame(data::AbstractArray{<:Real, 2}, coords::AbstractArray{<:Real, 2}, b::bbox, proj = 4326)

Usage

SpatialPixelsDataFrame(data, coords, bbox, proj)

Stars.bbox — Type

bbox(xmin, ymin, xmax, ymax)
bbox(;xmin, ymin, xmax, ymax)

Spatial bounding box

Stars.rast — Type

rast(fn::AbstractString, bands = nothing): construct GeoArray object from file

rast(ga::GeoArray; vals::AbstractArray{T}): Reconstruct GeoArray object

Index

Stars.GeoArray

Stars.SpatialPixelsDataFrame

Stars.bbox

Stars.rast

Base.getindex

Stars.cast_to_gdal

Stars.cellIJ

Stars.clip_point

Stars.equals

Stars.flipud!

Stars.gdal_polygonize

Stars.gdalinfo

Stars.get_nodata

Stars.maskCoords

Stars.mask_flags

Stars.readGDAL

Stars.readGDAL

Stars.shrink_bbox

Stars.st_as_sf

Stars.st_as_stars

Stars.st_bbox!

Stars.st_clip

Stars.st_coordinates

Stars.st_coordinates!

Stars.st_coords

Stars.st_crs!

Stars.st_dim

Stars.st_read

Stars.st_read

Stars.st_write

Stars.st_write